refactor(browser_exec): single shared dispatch, eliminates mcp/builtin drift (Closes #115)
Extracted all ~45 browser action arms that were duplicated between
mcp_server.rs and adk/tool/builtins.rs into a new src/browser_exec.rs
module with a single pub(crate) dispatch(action, input) entry point.

Both callers now delegate to browser_exec::dispatch(); caller-specific
logic is kept inline:
- builtins.rs: goto (authz check + screenshot event), screenshot
  (event publish), dom_snapshot (ref-id system), ax_tree (diff mode),
  screenshot_analyze (PNG capture sentinel for async LLM cache+SoM)
- mcp_server.rs: ext_status/url/tabs (Companion extension probes),
  screenshot_analyze (simple sync path), ocr_find_text

Actions gained by mcp_server from the merge: dom_snapshot, count,
value, select, scroll_to, hover, hover_point, new_tab, switch_tab,
close_tab, list_tabs, cookies, set_cookie, delete_cookie,
localStorage_get, localStorage_set, install_capture, viewport, pdf,
drag, http_auth, scroll (unified x/y params).

preflight.sh section 6 updated to search browser_exec.rs in addition
to builtins.rs so the action registry check remains green.
