+++ src/main.rs
@@ -8,6 +8,7 @@ pub mod error;
 mod platform;
 #[allow(dead_code)] mod browser;
 #[allow(dead_code)] mod browser_ax;
+#[allow(dead_code)] mod browser_exec;
 #[allow(dead_code)] mod claude_session;
 #[allow(dead_code)] mod multi_agent;
 #[allow(dead_code)] mod config_check;
